Factor compare_records into shared location
