+++ tests/conftest.py
@@ -9,6 +9,29 @@ TEST_ROOT = Path(__file__).resolve().parents[1]
 
 
 @pytest.fixture(scope='module')
+def compare_records():
+    """Provide compare_records method to test classes"""
+    def compare_records(actual_record: Dict, fixture_record: Dict):
+        """Check that identity records are identical."""
+        assert actual_record['concept_id'] == fixture_record['concept_id']
+        assert ('label' in actual_record) == ('label' in fixture_record)
+        if 'label' in actual_record or 'label' in fixture_record:
+            assert actual_record['label'] == fixture_record['label']
+        assert ('aliases' in actual_record) == ('aliases' in fixture_record)
+        if 'aliases' in actual_record or 'aliases' in fixture_record:
+            assert set(actual_record['aliases']) == set(fixture_record['aliases'])  # noqa: E501
+        print(actual_record.keys())
+        print(fixture_record.keys())
+        assert ('other_identifiers' in actual_record) == ('other_identifiers' in fixture_record)  # noqa: E501
+        if 'other_identifiers' in actual_record or 'other_identifiers' in fixture_record:  # noqa: E501
+            assert set(actual_record['other_identifiers']) == set(fixture_record['other_identifiers'])  # noqa: E501
+        assert ('xrefs' in actual_record) == ('xrefs' in fixture_record)
+        if 'xrefs' in actual_record or 'xrefs' in fixture_record:
+            assert set(actual_record['xrefs']) == set(fixture_record['xrefs'])
+    return compare_records
+
+
+@pytest.fixture(scope='module')
 def provide_root():
     """Provide TEST_ROOT value to test cases."""
     return TEST_ROOT
